Refactor backend and migraitons stack by extracting common resources to re-usable functions
