+++ packages/backend/infra/stacks/lib/names.ts
@@ -1,0 +1,5 @@
+import { EnvironmentSettings } from '@sb/infra-core';
+
+export function getBackendChamberServiceName(envSettings: EnvironmentSettings) {
+  return `env-${envSettings.projectEnvName}-backend`;
+}
